Conflicts:
	bin/classes/com/enge/bullethell/Components/Bullet_Component.class
	bin/classes/com/enge/bullethell/Components/Health_Component.class
	bin/classes/com/enge/bullethell/Components/Velocity_Component.class
	src/com/enge/bullethell/Components/Bullet_Component.java
	src/com/enge/bullethell/Components/Health_Component.java

+++ src/com/enge/bullethell/Components/Velocity_Component.java
@@ -4,16 +4,16 @@ import com.artemis.Component;
 
 public class Velocity_Component extends Component 
 {
-	public int velocity;
+	public float velocity;
 	
 	public Velocity_Component()
 	{
 		
 	}
 	
-	public Velocity_Component(int velocity)
+	public Velocity_Component(float velocity)
 	{
 		this.velocity = velocity;
 	}
-	
+
 }
